requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
m3u8downloader==0.10.1
markdown-it-py==3.0.0
mdurl==0.1.2
outcome==1.3.0.post0
Pygments==2.17.2
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
rich==13.7.1
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wells==1.5.0
wsproto==1.2.0
